+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.11.9] 2019-09-24
+
+- Add DSC resource for declaring poshbot configuration file.
+- Add example configuration for module, config file, and scheduled task
+
 ## [0.11.8] 2019-08-30
 
 ### Fixed